removes need for imports from __main__ module
creates src/nhp/model/run.py to prevent need to import from __main__

restructures tests folder to match the src folder
