+++ tests/nhp/model/test_health_status_adjustment.py
