+++ tests/nhp/model/test_outpatients.py
